Update
This version of the script processes all the files in the input directory in a single pass and stores the routes and their AS paths in a list called routes. After processing all the files, it creates the paths-dir directory (if it doesn't already exist) and writes the contents of routes to a single output file called routes.txt.