Update README.md
In the previous text, the "$" sign was included along with the command "pip install -r requirements.txt". However, the "$" sign is not part of the command and is used to represent the shell prompt in command line interfaces. To avoid confusion and ensure that users can copy the command without errors, the "$" sign has been removed from the text.